poc/digital_human_coach/pyproject.toml
[project]
name = "digital-human-coach"
version = "0.1.0"
description = "AI-powered Digital Human Communication Coach PoC"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "streamlit>=1.30.0",
    "gradio>=4.15.0",
    "openai>=1.10.0",
    "anthropic>=0.8.0",
    "deepgram-sdk>=3.2.0",
    "elevenlabs>=0.2.27",
    "edge-tts>=6.1.9",
    "gtts>=2.5.0",
    "openai-whisper>=20231117",
    "mediapipe>=0.10.9",
    "opencv-python>=4.9.0",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "numpy>=1.24.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
